requirements.txt
Flask
qrcode
pillow 
requests 
gunicorn 